Move `get_registered_details` fully into client
